snapshot/test_wal_validator.py
#!/usr/bin/env python3

#### META: Title: Wallace WAL File Validator Tests
#### META: Version: 0.1.1
#### META: Author: Claude + Human
#### META: PATH: tests/test_parser/test_wal_validator.py

#### SYNOPSIS: Test suite for .wal file validation component.
#### SYNOPSIS: Validates strict conformance to WAL spec 0.1.1.
#### SYNOPSIS: Tests both valid and invalid .wal structures.

#### CONTENTS:

import pytest
import json
from pathlib import Path

### SECTION: fixtures
@pytest.fixture
def valid_wal_json():
    """Create a valid .wal file in JSON format."""
    return {
        "wallace_version": "0.1.1",
        "file_type": "wal",
        "structure": {
            "shebang": "#!/usr/bin/env wallace",
            "header_block": {
                "meta_block": [
                    {"tag": "META", "key": "Title", "value": "Test WAL"},
                    {"tag": "META", "key": "Version", "value": "0.1.1"},
                    {"tag": "META", "key": "Author", "value": "Test"},
                    {"tag": "META", "key": "PATH", "value": "test.wal"}
                ],
                "synopsis_block": [
                    {"tag": "SYNOPSIS", "content": "Test WAL file."},
                    {"tag": "SYNOPSIS", "content": "For validation testing."}
                ],
                "contents_tag": {"tag": "CONTENTS"}
            },
            "contents_block": [
                {
                    "type": "section",
                    "name": "test",
                    "content": [
                        {"tag": "TODO", "content": "Implement test"},
                        {"tag": "METRICS", "values": [
                            {"value": 95, "unit": "%"},
                            {"value": 120, "unit": "mins"}
                        ]}
                    ],
                    "end": "test"
                }
            ]
        }
    }

@pytest.fixture
def invalid_wal_jsons():
    """Create various invalid .wal structures."""
    return {
        "missing_shebang": {
            "wallace_version": "0.1.1",
            "file_type": "wal",
            "structure": {
                "header_block": {},
                "contents_block": []
            }
        },
        "incomplete_meta": {
            "wallace_version": "0.1.1",
            "file_type": "wal",
            "structure": {
                "shebang": "#!/usr/bin/env wallace",
                "header_block": {
                    "meta_block": [
                        {"tag": "META", "key": "Title", "value": "Test"}
                        # Missing required META tags
                    ]
                }
            }
        },
        "missing_synopsis": {
            "wallace_version": "0.1.1",
            "file_type": "wal",
            "structure": {
                "shebang": "#!/usr/bin/env wallace",
                "header_block": {
                    "meta_block": []
                    # Missing synopsis_block
                }
            }
        },
        "unmatched_section": {
            "wallace_version": "0.1.1",
            "file_type": "wal",
            "structure": {
                "shebang": "#!/usr/bin/env wallace",
                "header_block": {},
                "contents_block": [
                    {
                        "type": "section",
                        "name": "test",
                        "content": [],
                        "end": "different"  # Mismatched section name
                    }
                ]
            }
        }
    }

### SECTION: structure_tests
def test_valid_wal_structure(valid_wal_json):
    """Test validation of correct WAL structure."""
    assert is_valid_wal(valid_wal_json) is True

def test_missing_shebang(invalid_wal_jsons):
    """Test detection of missing shebang."""
    assert is_valid_wal(invalid_wal_jsons["missing_shebang"]) is False

def test_incomplete_meta(invalid_wal_jsons):
    """Test detection of incomplete META block."""
    assert is_valid_wal(invalid_wal_jsons["incomplete_meta"]) is False

def test_missing_synopsis(invalid_wal_jsons):
    """Test detection of missing SYNOPSIS block."""
    assert is_valid_wal(invalid_wal_jsons["missing_synopsis"]) is False

def test_unmatched_section(invalid_wal_jsons):
    """Test detection of unmatched section tags."""
    assert is_valid_wal(invalid_wal_jsons["unmatched_section"]) is False

### SECTION: meta_tests
def test_required_meta_tags():
    """Test validation of required META tags."""
    # Must have Title, Version, Author, PATH

def test_meta_tag_order():
    """Test META tag order requirements."""
    # Order should not matter for validity

def test_meta_tag_uniqueness():
    """Test unique META tag requirements."""
    # Each key should appear only once

### SECTION: tag_tests
def test_valid_task_tags():
    """Test validation of task tags."""
    # TODO, DONE, FIXME, etc.

def test_valid_data_tags():
    """Test validation of data tags."""
    # METRICS, TAGS, FUNCTION, LINK

def test_valid_section_tags():
    """Test validation of section structure."""
    # SECTION/END pairs

### SECTION: metric_tests
def test_valid_metrics():
    """Test validation of METRICS tag values."""
    # Numbers with units

def test_invalid_metrics():
    """Test detection of invalid metrics."""
    # Wrong format, missing units

### SECTION: nesting_tests
def test_valid_nesting():
    """Test validation of nested sections."""
    # Proper nesting structure

def test_invalid_nesting():
    """Test detection of invalid nesting."""
    # Overlapping or mismatched sections

### METRICS: tests 15 coverage 100%

### FIXME: Need to verify exact METRICS format spec
### REVIEW: Consider strict vs lenient validation modes
